// src/app/models/permissions.ts
// Enum defining permissions for user actions and route access
export enum Permissions {
    CanCreateUser = 'CanCreateUser', // Permission to create a user
    CanReadUser = 'CanReadUser', // Permission to read user information
    CanUpdateUser = 'CanUpdateUser', // Permission to update user information
    CanDeleteUser = 'CanDeleteUser', // Permission to delete a user
    CanViewProtectedRoute1 = 'CanViewProtectedRoute1', // Permission to view ProtectedRoute1
    CanViewProtectedRoute2 = 'CanViewProtectedRoute2' // Permission to view ProtectedRoute2
}